주문 생성 API와 주문 결과 페이지 연동

+++ src/main/java/com/shop/domain/OrderItem.java
@@ -8,7 +8,7 @@ public class OrderItem {
 	private Long orderNo;            // 주문번호
 	private Long productOptionNo;    // 상품옵션번호
 	private Integer quantity;        // 수량
-	private Integer unitPrice;       // 단가
+	private Long unitPrice;       // 단가
 	private String itemName;         // 상품명
 	private String itemSize;         // 상품사이즈
 	private String itemColor;        // 상품색상
